Refactoring of middleware and tests
