+++ .env
@@ -1,7 +1,6 @@
 DB_USER=public
 DB_PASS=123
 PORT=8081
-DB=localhost
-DEBUG=YES
-HEROKU=NO
+DB=localhosts
 SECRET=EnCry&ting_InPr0gre&#+Do-N@t.mE&S=it|Up,Th%nK$.
+ACCESS_ALLOW_ORIGIN=*
