+++ middleware/auth.js
@@ -43,8 +43,7 @@ var middleware = {
         });
     }
 };
-middleware.canEditLevel = function (req, res, next) {
-
+function ownsLevelOrAdminPowers(req, res, adminPowers, next) {
     middleware.loggedIn(req, res, function () {
         Level.findById(req.params.levelid, function (err, foundLevel) {
             if (err) {
@@ -52,7 +51,7 @@ middleware.canEditLevel = function (req, res, next) {
             } else if (!foundLevel) {
                 sendJSON(res, "error", { message: "No level found, it probably doesn't exist" }, 400);
             } else {
-                if (req.user._id.equals(foundLevel.creator.id) || req.user.adminPowers >= 1) {
+                if (req.user._id.equals(foundLevel.creator.id) || req.user.adminPowers >= adminPowers) {
                     next();
                 } else {
                     sendJSON(res, "error", { message: "Make sure you own the level first", error: "Insufficient Permissions" }, 400);
@@ -61,11 +60,18 @@ middleware.canEditLevel = function (req, res, next) {
         });
     });
 }
-// middleware.canDeleteLevel = function (req, res, next) {
-
-// };
-// middleware.isAdmin = function(req, res, next) {
-
-// }
+middleware.canEditLevel = function (req, res, next) {
+    ownsLevelOrAdminPowers(req, res, 2, next);
+}
+middleware.canDeleteLevel = function (req, res, next) {
+    ownsLevelOrAdminPowers(req, res, 1, next);
+};
+middleware.isAdmin = function(req, res, next) {
+    if(req.user.adminPowers >= 2) {
+        next();
+    } else {
+        sendJSON(res, "error", { message: "Not an admin!", error: "Insufficient Permissions" }, 400);
+    }
+}
 
 module.exports = middleware;
