+++ tests/data/fail-levels.js
@@ -1,0 +1,13 @@
+
+// A bunch of levels that by failing pass
+var typeFailLevel = {
+    title: "Type Not Valid Test",
+    type: "non-valid so fail",
+    levelData: {
+        "a bunch of random stuff": "that doesnt work"
+    },
+    difficulty: 100000
+}
+// var 
+
+module.exports = [typeFailLevel];
